Move ssl protobuf reader to message_translation
