+++ src/software/proto/message_translation/ssl_wrapper.h
