+++ src/software/proto/message_translation/tbots_protobuf_test.cpp
@@ -1,4 +1,4 @@
-#include "software/proto/message_translation/tbots_protobuf_message_translation.h"
+#include "software/proto/message_translation/tbots_protobuf.h"
 
 #include <gtest/gtest.h>
 #include <string.h>
