+++ src/software/sensor_fusion/sensor_fusion.h
@@ -5,11 +5,11 @@
 #include "software/backend/robot_status.h"
 #include "software/multithreading/subject.h"
 #include "software/multithreading/threaded_observer.h"
+#include "software/proto/message_translation/ssl_protobuf_reader.h"
 #include "software/proto/sensor_msg.pb.h"
 #include "software/sensor_fusion/filter/ball_filter.h"
 #include "software/sensor_fusion/filter/robot_team_filter.h"
 #include "software/sensor_fusion/refbox_data.h"
-#include "software/sensor_fusion/ssl_protobuf_reader.h"
 #include "software/sensor_fusion/vision_detection.h"
 #include "software/world/ball.h"
 #include "software/world/team.h"
