+++ src/software/simulation/BUILD
@@ -126,9 +126,9 @@ cc_library(
         "//firmware/app/world:firmware_world",
         "//software/backend/output/radio/mrf:mrf_primitive_visitor",
         "//software/primitive",
-        "//software/proto/message_translation:ssl_detection_message_translator",
-        "//software/proto/message_translation:ssl_geometry_message_translator",
-        "//software/proto/message_translation:ssl_wrapper_message_translator",
+        "//software/proto/message_translation:ssl_detection",
+        "//software/proto/message_translation:ssl_geometry",
+        "//software/proto/message_translation:ssl_wrapper",
         "//software/simulation/physics:physics_world",
         "//software/world",
     ],
